requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
starlette==0.37.2
PyYAML==6.0.2
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.9.0.80
# ML
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
diffusers>=0.27.2
transformers>=4.42.0
accelerate>=0.30.0
safetensors>=0.4.3
xformers>=0.0.26.post1 ; platform_system=="Windows"
# Pose / segmentation (lightweight defaults)
mediapipe>=0.10.14
# Optional / video
imageio[ffmpeg]>=2.34.0
moviepy>=1.0.3
# Optional lip-sync (Wav2Lip) & audio
librosa>=0.10.2.post1
